drop “rebootable” compat
now, we can move it into the correct directory
