+++ lib/vagrant-vbguest/helpers/rebootable.rb
@@ -1,9 +1,31 @@
-require 'vagrant-vbguest/rebootable'
-
 module VagrantVbguest
   module Helpers
-
     module Rebootable
+      include VmCompatible
+
+      def self.included(base)
+        base.extend(ClassMethods)
+      end
+
+      @@rebooted = {}
+
+      def rebooted?(vm)
+        !!@@rebooted[ self.class.vm_id(vm) ]
+      end
+
+      def reboot?(vm, options)
+        if rebooted?(vm)
+          vm.env.ui.error(I18n.t("vagrant_vbguest.restart_loop_guard_activated"))
+          false
+        elsif options[:auto_reboot]
+          vm.env.ui.warn(I18n.t("vagrant_vbguest.restart_vm"))
+          @@rebooted[ self.class.vm_id(vm) ] = true
+        else
+          vm.env.ui.warn(I18n.t("vagrant_vbguest.suggest_restart", :name => vm.name))
+          false
+        end
+      end
+
       def reboot(vm, options)
         if reboot? vm, options
           simple_reboot = Vagrant::Action::Builder.new.tap do |b|
@@ -28,6 +50,5 @@ module VagrantVbguest
         end
       end
     end
-
   end
 end
