--- lib/vagrant-vbguest/vagrant_compat/vagrant_1_2/rebootable.rb
@@ -1,1 +1,0 @@
-require File.expand_path("../../vagrant_1_1/rebootable", __FILE__)
